Display Options (#359)
Implemented render mode option and switched html background color to black.

"Pixel Imperfect" render mode is the new default and users can switch to the old "Pixel Perfect" mode if desired. A third "Native" mode is available in Dev mode and is forced to be used when opening the dev zoom map (tab key).